moving common functions into util, and move quick pick code
