+++ src/commands/aksAccount/aksAccount.ts
@@ -1,10 +1,9 @@
-import { QuickPickItem, Uri, env, window } from "vscode";
+import { window } from "vscode";
 import { failed } from "../utils/errorable";
-import { SubscriptionFilter, getFilteredSubscriptions, setFilteredSubscriptions } from "../utils/config";
+import { getFilteredSubscriptions, setFilteredSubscriptions } from "../utils/config";
 import { getSessionProvider } from "../../auth/azureSessionProvider";
-import { DefinedSubscription, SelectionType, getSubscriptions } from "../utils/subscriptions";
+import { SelectionType, SubscriptionQuickPickItem, getSubscriptions, handleNoSubscriptionsFound } from "../utils/subscriptions";
 import { getReadySessionProvider, quickPickTenant } from "../../auth/azureAuth";
-import { ReadyAzureSessionProvider } from "../../auth/types";
 
 export async function signInToAzure(): Promise<void> {
     await getSessionProvider().signIn();
@@ -40,7 +39,6 @@ export async function selectTenant(): Promise<void> {
     sessionProvider.selectedTenant = selectedTenant;
 }
 
-type SubscriptionQuickPickItem = QuickPickItem & { subscription: SubscriptionFilter };
 
 export async function selectSubscriptions(): Promise<void> {
     const sessionProvider = await getReadySessionProvider();
@@ -49,10 +47,13 @@ export async function selectSubscriptions(): Promise<void> {
         return;
     }
 
-    const allSubscriptions = await getAllSubscriptions(sessionProvider.result);
-    if (!allSubscriptions) return;
+    const allSubscriptions = await getSubscriptions(sessionProvider.result, SelectionType.All);
+    if (failed(allSubscriptions)) {
+        await window.showErrorMessage(allSubscriptions.error);
+        return;
+    }
 
-    if (allSubscriptions.length === 0) {
+    if (allSubscriptions.result.length === 0) {
         await handleNoSubscriptionsFound();
         return;
     }
@@ -70,7 +71,7 @@ export async function selectSubscriptions(): Promise<void> {
     );
     const subscriptionsInOtherTenants = filteredSubscriptions.filter((sub) => sub.tenantId !== session.result.tenantId);
 
-    const quickPickItems: SubscriptionQuickPickItem[] = allSubscriptions.map((sub) => {
+    const quickPickItems: SubscriptionQuickPickItem[] = allSubscriptions.result.map((sub) => {
         return {
             label: sub.displayName || "",
             description: sub.subscriptionId,
@@ -100,60 +101,4 @@ export async function selectSubscriptions(): Promise<void> {
     ];
 
     await setFilteredSubscriptions(newFilteredSubscriptions);
-}
-
-export async function selectSubscription(): Promise<string | undefined> {
-    const sessionProvider = await getReadySessionProvider();
-    if (failed(sessionProvider)) {
-        window.showErrorMessage(sessionProvider.error);
-        return;
-    }
-
-    const allSubscriptions = await getAllSubscriptions(sessionProvider.result);
-    if (!allSubscriptions) return;
-
-    if (allSubscriptions.length === 0) {
-        await handleNoSubscriptionsFound();
-        return;
-    }
-
-    const quickPickItems: SubscriptionQuickPickItem[] = allSubscriptions.map((sub) => {
-        return {
-            label: sub.displayName || "",
-            description: sub.subscriptionId,
-            subscription: {
-                subscriptionId: sub.subscriptionId || "",
-                tenantId: sub.tenantId || "",
-            },
-        };
-    });
-
-    const selectedItem = await window.showQuickPick(quickPickItems, {
-        canPickMany: false,
-        placeHolder: "Select a Subscription",
-    });
-
-    if(!selectedItem) {
-        return undefined;
-    }
-
-    return selectedItem.subscription.subscriptionId;
-}
-
-async function getAllSubscriptions(sessionProvider: ReadyAzureSessionProvider): Promise<DefinedSubscription[] | null> {
-    const allSubscriptions = await getSubscriptions(sessionProvider, SelectionType.All);
-    if (failed(allSubscriptions)) {
-        await window.showErrorMessage(allSubscriptions.error);
-        return null;
-    }
-    return allSubscriptions.result;
-}
-
-async function handleNoSubscriptionsFound(): Promise<void> {
-    const noSubscriptionsFound = "No subscriptions were found. Set up your account if you have yet to do so.";
-    const setupAccount = "Set up Account";
-    const response = await window.showInformationMessage(noSubscriptionsFound, setupAccount);
-    if (response === setupAccount) {
-        env.openExternal(Uri.parse("https://azure.microsoft.com/"));
-    }
 }
